Add EmailTagHelper for custom <kontakt> tag in Razor views
Introduce a new EmailTagHelper class to the ASP.NET Core MVC project, enabling the use of a custom <kontakt> tag in Razor views. This tag helper converts the <kontakt> tag into an anchor (<a>) tag with a mailto: link. The DisplayName and Adresse properties set the display text and email address of the link, respectively. Updated _Layout.cshtml to use the new <kontakt> tag for contact information in the footer.