Add GetAuthorById endpoint and update AddAuthor response
Introduced a new endpoint in `AuthorsController` to get an author by their ID using the `GetAuthorByIdAsync` method. This method retrieves an author by their `Guid` and returns appropriate HTTP responses.

Uncommented the `CreatedAtRoute` line in `AddAuthorAsync` to include a route to the newly created author in the response. Commented out the previously active `return Created();` line as it is no longer needed.